Run the UI in process and replace (some) sh with Python
Running the UI in-process guarantees that a UI crash will stop the flow
of video information.  A UI crash that did not stop streaming would be a
security risk.  Python code is both more flexible and easier to secure
than shell scripts.

The new code uses a “Service” class to encapsulate the common code
shared between both ‘qvc.Webcam’ and ‘qvc.ScreenShare’.  Only
per-service code needs to be separate.