Compatiblity fixes for Qubes R4.0 stable
All but one of these fixes are for the video sending component so it
works in Qubes R4.0 dom0 (Fedora 25). This means support for Python
3.5.4 and GStreamer 1.10.5. We will be able to drop these fixes when
Qubes R4.1 is released giving us a more up-to-date dom0.

Note that even if ones webcam is in sys-usb, compatibility with dom0
would still be necessary for screen sharing the host screen. This will
remain true until we have achieved GUIVM.

Additionally, support an earlier version of qrexec-client-vm on the
video receiving side on Qubes R4.0 AppVMs.